Apply new WFH for 2026+
