+++ Database/tg/fortran/wfhmodule/chidist.csv
@@ -1,0 +1,7 @@
+wfhgroup,hhchi2,pct
+high,1,0.405
+high,2,0.595
+medium,1,0.354
+medium,2,0.646
+low,1,0.37
+low,2,0.63
