+++ Database/tg/fortran/wfhmodule/edudist.csv
@@ -1,0 +1,7 @@
+wfhgroup,hhedu2,pct
+high,1,0.202
+high,2,0.798
+medium,1,0.268
+medium,2,0.732
+low,1,0.305
+low,2,0.695
