+++ Database/tg/fortran/wfhmodule/incdist.csv
@@ -1,9 +1,7 @@
-incdist,hhinc4,pct
-high,1,0.038985164
-high,2,0.080435697
-high,3,0.156394004
-high,4,0.724185136
-low,1,0.09287371
-low,2,0.134953622
-low,3,0.238871815
-low,4,0.533300852
+wfhgroup,hhinc2,pct
+high,1,0.231
+high,2,0.769
+medium,1,0.418
+medium,2,0.582
+low,1,0.353
+low,2,0.647
