+++ Database/tg/fortran/wfhmodule/indp_naics.csv
@@ -1,295 +1,295 @@
-INDP,naics2
-170,11
-180,11
-190,11
-270,11
-280,11
-290,11
-370,21
-380,21
-390,21
-470,21
-490,21
-570,22
-580,22
-590,22
-670,22
-680,22
-690,22
-770,23
-1070,31-33
-1080,31-33
-1090,31-33
-1170,31-33
-1180,31-33
-1190,31-33
-1270,31-33
-1280,31-33
-1290,31-33
-1370,31-33
-1390,31-33
-1470,31-33
-1480,31-33
-1490,31-33
-1570,31-33
-1590,31-33
-1670,31-33
-1680,31-33
-1690,31-33
-1691,31-33
-1770,31-33
-1790,31-33
-1870,31-33
-1880,31-33
-1890,31-33
-1990,31-33
-2070,31-33
-2090,31-33
-2170,31-33
-2180,31-33
-2190,31-33
-2270,31-33
-2280,31-33
-2290,31-33
-2370,31-33
-2380,31-33
-2390,31-33
-2470,31-33
-2480,31-33
-2490,31-33
-2570,31-33
-2590,31-33
-2670,31-33
-2680,31-33
-2690,31-33
-2770,31-33
-2780,31-33
-2790,31-33
-2870,31-33
-2880,31-33
-2890,31-33
-2970,31-33
-2980,31-33
-2990,31-33
-3070,31-33
-3080,31-33
-3090,31-33
-3095,31-33
-3170,31-33
-3180,31-33
-3190,31-33
-3291,31-33
-3360,31-33
-3365,31-33
-3370,31-33
-3380,31-33
-3390,31-33
-3470,31-33
-3490,31-33
-3570,31-33
-3580,31-33
-3590,31-33
-3670,31-33
-3680,31-33
-3690,31-33
-3770,31-33
-3780,31-33
-3790,31-33
-3870,31-33
-3875,31-33
-3890,31-33
-3895,31-33
-3960,31-33
-3970,31-33
-3980,31-33
-3990,31-33
-4070,42
-4080,42
-4090,42
-4170,42
-4180,42
-4190,42
-4195,42
-4260,42
-4265,42
-4270,42
-4280,42
-4290,42
-4370,42
-4380,42
-4390,42
-4470,42
-4480,42
-4490,42
-4560,42
-4570,42
-4580,42
-4585,42
-4590,42
-4670,44-45
-4680,44-45
-4690,44-45
-4770,44-45
-4780,44-45
-4790,44-45
-4795,44-45
-4870,44-45
-4880,44-45
-4890,44-45
-4970,44-45
-4971,44-45
-4972,44-45
-4980,44-45
-4990,44-45
-5070,44-45
-5080,44-45
-5090,44-45
-5170,44-45
-5180,44-45
-5190,44-45
-5270,44-45
-5275,44-45
-5280,44-45
-5290,44-45
-5295,44-45
-5370,44-45
-5380,44-45
-5381,44-45
-5390,44-45
-5391,44-45
-5470,44-45
-5480,44-45
-5490,44-45
-5570,44-45
-5580,44-45
-5590,44-45
-5591,44-45
-5592,44-45
-5593,44-45
-5670,44-45
-5680,44-45
-5690,44-45
-5790,44-45
-6070,48-49
-6080,48-49
-6090,48-49
-6170,48-49
-6180,48-49
-6190,48-49
-6270,48-49
-6280,48-49
-6290,48-49
-6370,48-49
-6380,48-49
-6390,48-49
-6470,51
-6480,51
-6490,51
-6570,51
-6590,51
-6670,51
-6672,51
-6680,51
-6690,51
-6695,51
-6770,51
-6780,51
-6870,52
-6880,52
-6890,52
-6970,52
-6990,52
-6991,52
-6992,52
-7070,53
-7071,53
-7072,53
-7080,53
-7170,53
-7180,53
-7181,53
-7190,53
-7270,54
-7280,54
-7290,54
-7370,54
-7380,54
-7390,54
-7460,54
-7470,54
-7480,54
-7490,54
-7570,55
-7580,56
-7590,56
-7670,56
-7680,56
-7690,56
-7770,56
-7780,56
-7790,56
-7860,61
-7870,61
-7880,61
-7890,61
-7970,62
-7980,62
-7990,62
-8070,62
-8080,62
-8090,62
-8170,62
-8180,62
-8190,62
-8191,62
-8192,62
-8270,62
-8290,62
-8370,62
-8380,62
-8390,62
-8470,62
-8560,71
-8561,71
-8562,71
-8563,71
-8564,71
-8570,71
-8580,71
-8590,71
-8660,72
-8670,72
-8680,72
-8690,72
-8770,81
-8780,81
-8790,81
-8870,81
-8880,81
-8891,81
-8970,81
-8980,81
-8990,81
-9070,81
-9080,81
-9090,81
-9160,81
-9170,81
-9180,81
-9190,81
-9290,81
-9370,92
-9380,92
-9390,92
-9470,92
-9480,92
-9490,92
-9570,92
-9590,92
-9670,92
-9680,92
-9690,92
-9770,92
-9780,92
-9790,92
-9870,92
+INDP,naics2,trc
+170,11,low
+180,11,low
+190,11,low
+270,11,low
+280,11,low
+290,11,low
+370,21,low
+380,21,low
+390,21,low
+470,21,low
+490,21,low
+570,22,medium
+580,22,medium
+590,22,medium
+670,22,medium
+680,22,medium
+690,22,medium
+770,23,low
+1070,31-33,low
+1080,31-33,low
+1090,31-33,low
+1170,31-33,low
+1180,31-33,low
+1190,31-33,low
+1270,31-33,low
+1280,31-33,low
+1290,31-33,low
+1370,31-33,low
+1390,31-33,low
+1470,31-33,low
+1480,31-33,low
+1490,31-33,low
+1570,31-33,low
+1590,31-33,low
+1670,31-33,low
+1680,31-33,low
+1690,31-33,low
+1691,31-33,low
+1770,31-33,low
+1790,31-33,low
+1870,31-33,low
+1880,31-33,low
+1890,31-33,low
+1990,31-33,low
+2070,31-33,low
+2090,31-33,low
+2170,31-33,low
+2180,31-33,low
+2190,31-33,low
+2270,31-33,low
+2280,31-33,low
+2290,31-33,low
+2370,31-33,low
+2380,31-33,low
+2390,31-33,low
+2470,31-33,low
+2480,31-33,low
+2490,31-33,low
+2570,31-33,low
+2590,31-33,low
+2670,31-33,low
+2680,31-33,low
+2690,31-33,low
+2770,31-33,low
+2780,31-33,low
+2790,31-33,low
+2870,31-33,low
+2880,31-33,low
+2890,31-33,low
+2970,31-33,low
+2980,31-33,low
+2990,31-33,low
+3070,31-33,low
+3080,31-33,low
+3090,31-33,low
+3095,31-33,low
+3170,31-33,low
+3180,31-33,low
+3190,31-33,low
+3291,31-33,low
+3360,31-33,low
+3365,31-33,low
+3370,31-33,low
+3380,31-33,low
+3390,31-33,low
+3470,31-33,low
+3490,31-33,low
+3570,31-33,low
+3580,31-33,low
+3590,31-33,low
+3670,31-33,low
+3680,31-33,low
+3690,31-33,low
+3770,31-33,low
+3780,31-33,low
+3790,31-33,low
+3870,31-33,low
+3875,31-33,low
+3890,31-33,low
+3895,31-33,low
+3960,31-33,low
+3970,31-33,low
+3980,31-33,low
+3990,31-33,low
+4070,42,medium
+4080,42,medium
+4090,42,medium
+4170,42,medium
+4180,42,medium
+4190,42,medium
+4195,42,medium
+4260,42,medium
+4265,42,medium
+4270,42,medium
+4280,42,medium
+4290,42,medium
+4370,42,medium
+4380,42,medium
+4390,42,medium
+4470,42,medium
+4480,42,medium
+4490,42,medium
+4560,42,medium
+4570,42,medium
+4580,42,medium
+4585,42,medium
+4590,42,medium
+4670,44-45,low
+4680,44-45,low
+4690,44-45,low
+4770,44-45,low
+4780,44-45,low
+4790,44-45,low
+4795,44-45,low
+4870,44-45,low
+4880,44-45,low
+4890,44-45,low
+4970,44-45,low
+4971,44-45,low
+4972,44-45,low
+4980,44-45,low
+4990,44-45,low
+5070,44-45,low
+5080,44-45,low
+5090,44-45,low
+5170,44-45,low
+5180,44-45,low
+5190,44-45,low
+5270,44-45,low
+5275,44-45,low
+5280,44-45,low
+5290,44-45,low
+5295,44-45,low
+5370,44-45,low
+5380,44-45,low
+5381,44-45,low
+5390,44-45,low
+5391,44-45,low
+5470,44-45,low
+5480,44-45,low
+5490,44-45,low
+5570,44-45,low
+5580,44-45,low
+5590,44-45,low
+5591,44-45,low
+5592,44-45,low
+5593,44-45,low
+5670,44-45,low
+5680,44-45,low
+5690,44-45,low
+5790,44-45,low
+6070,48-49,low
+6080,48-49,low
+6090,48-49,low
+6170,48-49,low
+6180,48-49,low
+6190,48-49,low
+6270,48-49,low
+6280,48-49,low
+6290,48-49,low
+6370,48-49,low
+6380,48-49,low
+6390,48-49,low
+6470,51,high
+6480,51,high
+6490,51,high
+6570,51,high
+6590,51,high
+6670,51,high
+6672,51,high
+6680,51,high
+6690,51,high
+6695,51,high
+6770,51,high
+6780,51,high
+6870,52,high
+6880,52,high
+6890,52,high
+6970,52,high
+6990,52,high
+6991,52,high
+6992,52,high
+7070,53,high
+7071,53,high
+7072,53,high
+7080,53,high
+7170,53,high
+7180,53,high
+7181,53,high
+7190,53,high
+7270,54,high
+7280,54,high
+7290,54,high
+7370,54,high
+7380,54,high
+7390,54,high
+7460,54,high
+7470,54,high
+7480,54,high
+7490,54,high
+7570,55,low
+7580,56,low
+7590,56,low
+7670,56,low
+7680,56,low
+7690,56,low
+7770,56,low
+7780,56,low
+7790,56,low
+7860,61,low
+7870,61,low
+7880,61,low
+7890,61,low
+7970,62,low
+7980,62,low
+7990,62,low
+8070,62,low
+8080,62,low
+8090,62,low
+8170,62,low
+8180,62,low
+8190,62,low
+8191,62,low
+8192,62,low
+8270,62,low
+8290,62,low
+8370,62,low
+8380,62,low
+8390,62,low
+8470,62,low
+8560,71,medium
+8561,71,medium
+8562,71,medium
+8563,71,medium
+8564,71,medium
+8570,71,medium
+8580,71,medium
+8590,71,medium
+8660,72,low
+8670,72,low
+8680,72,low
+8690,72,low
+8770,81,low
+8780,81,low
+8790,81,low
+8870,81,low
+8880,81,low
+8891,81,low
+8970,81,low
+8980,81,low
+8990,81,low
+9070,81,low
+9080,81,low
+9090,81,low
+9160,81,low
+9170,81,low
+9180,81,low
+9190,81,low
+9290,81,low
+9370,92,medium
+9380,92,medium
+9390,92,medium
+9470,92,medium
+9480,92,medium
+9490,92,medium
+9570,92,medium
+9590,92,medium
+9670,92,medium
+9680,92,medium
+9690,92,medium
+9770,92,medium
+9780,92,medium
+9790,92,medium
+9870,92,medium
